Reuse pltform_name() function
