+++ audresample/core/lib.py
@@ -3,29 +3,62 @@ import os
 import platform
 
 
+def platform_name():
+
+    # Look for enrionment variable
+    # to be able to enforce
+    # different platform names
+    # in CI on the same runner
+    plat_name = os.environ.get('PLAT_NAME', None)
+
+    if plat_name is None:
+        # Extract platform name from system + processor
+        system = platform.system()
+        processor = platform.processor()
+
+        if system == 'Linux' and processor == 'x86_64':
+            # That we support 2_17 can be seen
+            # when inspecting the wheel with auditwheel
+            plat_name = 'manylinux_2_17_x86_64'
+
+        elif system == 'Windows':
+            plat_name = 'win_amd64'
+
+        elif system == 'Darwin' and processor == 'i386':
+            plat_name = 'macosx_12_0_x86_64'
+
+        elif system == 'Darwin' and processor == 'arm':
+            plat_name = 'macosx_12_0_arm64'
+
+        else:
+            raise RuntimeError(
+                f'Unsupported platform {system}-{processor}'
+            )
+
+    return plat_name
+
+
 # load library
 
 root = os.path.dirname(os.path.realpath(__file__))
-
+plat_name = platform_name()
 bin_path = os.path.join(root, 'bin')
-if platform.system() == 'Windows':  # pragma: no cover
+
+if plat_name == 'win_amd64':  # pragma: no cover
     lib_path = os.path.join(bin_path, 'windows', 'audresample.dll')
-elif platform.system() == 'Linux':  # pragma: no cover
+
+elif plat_name == 'manylinux_2_17_x86_64':  # pragma: no cover
     lib_path = os.path.join(bin_path, 'linux', 'libaudresample.so')
-elif (
-        # MacOS Intel
-        platform.system() == 'Darwin'
-        and platform.processor() == 'i386'
-):  # pragma: no cover
+
+elif plat_name == 'macosx_12_0_x86_64':  # pragma: no cover
     lib_path = os.path.join(bin_path, 'macos-intel', 'libaudresample.dylib')
-elif (
-        # MacOS M1
-        platform.system() == 'Darwin'
-        and platform.processor() == 'arm'
-):  # pragma: no cover
+
+elif plat_name == 'macosx_12_0_arm64':  # pragma: no cover
     lib_path = os.path.join(bin_path, 'macos-m1', 'libaudresample.dylib')
+
 else:  # pragma: no cover
     raise RuntimeError("Unsupported platform")
+
 lib = ctypes.cdll.LoadLibrary(lib_path)
